backend/requirements.txt
fastapi
uvicorn
qdrant-client
redis
torch
diffusers
transformers
accelerate
numpy
opencv-python
pillow
SpeechRecognition
pyaudio
controlnet-aux
deepface
python-multipart
